requirements.txt
Pillow
imagehash
opencv-python
numpy
matplotlib
scikit-image
torch
torchvision
facenet-pytorch 